+++ soul-admin/src/main/java/org/dromara/soul/admin/disruptor/SoulServerMetaDataRegisterEventPublisher.java
@@ -17,18 +17,30 @@
 
 package org.dromara.soul.admin.disruptor;
 
+import org.dromara.soul.admin.dto.MetaDataDTO;
+import org.dromara.soul.admin.dto.SpringCloudRegisterDTO;
+import org.dromara.soul.admin.dto.SpringMvcRegisterDTO;
+import org.dromara.soul.admin.mapper.SelectorMapper;
+import org.dromara.soul.admin.service.SelectorService;
 import org.dromara.soul.admin.service.SoulClientRegisterService;
+import org.dromara.soul.common.enums.RpcTypeEnum;
+import org.dromara.soul.common.utils.GsonUtils;
 import org.dromara.soul.disruptor.DisruptorProviderManage;
+import org.dromara.soul.register.server.api.SoulSeverRegisterCenterEventPublisher;
+import org.dromara.soul.register.server.api.listener.DataChangedEvent;
+import org.springframework.context.ApplicationEventPublisher;
 
 /**
  * SoulServerMetaDataRegisterEventPublisher.
  *
  * @author tydhot
+ * @author lw1243925457
  */
-public class SoulServerMetaDataRegisterEventPublisher {
+public class SoulServerMetaDataRegisterEventPublisher implements SoulSeverRegisterCenterEventPublisher {
+
     private static final SoulServerMetaDataRegisterEventPublisher INSTANCE = new SoulServerMetaDataRegisterEventPublisher();
 
-    private DisruptorProviderManage<SoulServerMetaDataRegisterEvent> disruptorProviderManage;
+    private DisruptorProviderManage<DataChangedEvent> disruptorProviderManage;
 
     /**
      * get instance.
@@ -44,27 +56,55 @@ public class SoulServerMetaDataRegisterEventPublisher {
      *
      * @param soulClientRegisterService soulClientRegisterService
      */
-    public void start(final SoulClientRegisterService soulClientRegisterService) {
-        disruptorProviderManage =
-                new DisruptorProviderManage<>(
-                        new SoulServerMetaDataRegisterEventHandler(soulClientRegisterService), 1, 4096 * 2 * 2);
+    public void start(final SoulClientRegisterService soulClientRegisterService,
+                      final SelectorService selectorService,
+                      final SelectorMapper selectorMapper,
+                      final ApplicationEventPublisher eventPublisher) {
+        disruptorProviderManage = new DisruptorProviderManage<>(
+                new SoulServerMetaDataRegisterEventHandler(soulClientRegisterService, selectorService, selectorMapper, eventPublisher),
+                1, 4096 * 2 * 2);
         disruptorProviderManage.startup();
     }
 
     /**
+     * 注册中心应该只放入数据，数据应该在Admin中，那事件应该有多种类型：注册事件、更新事件（服务下线）
+     * 所以这里增加了一个更新事件
+     *
      * publish event.
      *
-     * @param metaDataDTO metaDataDTO
-     * @param rpcType rpcType
+     * @param eventType data event type
+     * @param key if register event, key is eventType; if update event, key is context path
+     * @param value if register event, value is metadata; if update event, key is uri list
      */
-    public void publishEvent(final String rpcType, final Object metaDataDTO) {
-        SoulServerMetaDataRegisterEvent event = new SoulServerMetaDataRegisterEvent();
-        event.setMetadata(metaDataDTO);
-        event.setType(rpcType);
+    @Override
+    public void publishEvent(final DataChangedEvent.Type eventType, final String key, final Object value) {
+        // TODO:两种事件混用，参数和处理之类，给人感觉不够清晰明了
+        DataChangedEvent event;
+        if (eventType.equals(DataChangedEvent.Type.REGISTER)) {
+            // 为了兼容HTTP注册方式：目前HTTP能直接得到Admin中的数据类型
+            // 但ZK Register模块不应该依赖于Admin相关模块；因为ZK数据以字符串形式传入，所以在这里进行数据转换
+            if (value instanceof String) {
+                event = new DataChangedEvent(key, convertToDTO(key, value.toString()), eventType);
+            } else {
+                event = new DataChangedEvent(key, value, eventType);
+            }
+        } else {
+            event = new DataChangedEvent(key, value, eventType);
+        }
         push(event);
     }
 
-    private void push(final SoulServerMetaDataRegisterEvent event) {
+    private Object convertToDTO(String type, String data) {
+        if (type.equals(RpcTypeEnum.HTTP.getName())) {
+            return GsonUtils.getInstance().fromJson(data, SpringMvcRegisterDTO.class);
+        }
+        if (type.equals(RpcTypeEnum.SPRING_CLOUD.getName())) {
+            return GsonUtils.getInstance().fromJson(data, SpringCloudRegisterDTO.class);
+        }
+        return GsonUtils.getInstance().fromJson(data, MetaDataDTO.class);
+    }
+
+    private void push(final DataChangedEvent event) {
         disruptorProviderManage.getProvider().onData(event);
     }
 }
