+++ soul-examples/soul-examples-dubbo/soul-examples-alibaba-dubbo-service/src/main/resources/application.yml
@@ -23,8 +23,10 @@ spring:
     allow-bean-definition-overriding: true
 soul:
   client:
-    registerType: http
-    serverLists: http://localhost:9095
+#    registerType: http
+#    serverLists: http://localhost:9095
+    registerType: zookeeper
+    serverLists: localhost:2181
     props:
       contextPath: /dubbo
       appName: dubbo
