+++ soul-examples/soul-examples-dubbo/soul-examples-apache-dubbo-service/src/main/resources/application.yml
@@ -24,8 +24,10 @@ spring:
 
 soul:
   client:
-    registerType: http
-    serverLists: http://localhost:9095
+#    registerType: http
+#    serverLists: http://localhost:9095
+    registerType: zookeeper
+    serverLists: localhost:2181
     props:
       contextPath: /dubbo
       appName: dubbo
