+++ soul-examples/soul-examples-http/src/main/resources/application.yml
@@ -20,10 +20,10 @@ server:
 
 soul:
   client:
-    registerType: http
-    serverLists: http://localhost:9095
-#    registerType: zookeeper
-#    serverLists: localhost:2181
+#    registerType: http
+#    serverLists: http://localhost:9095
+    registerType: zookeeper
+    serverLists: localhost:2181
     props:
       contextPath: /http
       appName: http
