+++ soul-examples/soul-examples-sofa/src/main/resources/application.yml
@@ -32,8 +32,10 @@ com:
 
 soul:
   client:
-    registerType: http
-    serverLists: http://localhost:9095
+#    registerType: http
+#    serverLists: http://localhost:9095
+    registerType: zookeeper
+    serverLists: localhost:2181
     props:
       contextPath: /sofa
       ipAndPort: sofa
