+++ soul-examples/soul-examples-springcloud/src/main/resources/application.yml
@@ -36,8 +36,10 @@ eureka:
 
 soul:
   client:
-    registerType: http
-    serverLists: http://localhost:9095
+#    registerType: http
+#    serverLists: http://localhost:9095
+    registerType: zookeeper
+    serverLists: localhost:2181
     props:
       contextPath: /springcloud
