+++ soul-examples/soul-examples-tars/src/main/resources/application.yml
@@ -30,8 +30,10 @@ logging:
 
 soul:
   client:
-    registerType: http
-    serverLists: http://localhost:9095
+#    registerType: http
+#    serverLists: http://localhost:9095
+    registerType: zookeeper
+    serverLists: localhost:2181
     props:
       contextPath: /tars
       appName: tars
