+++ soul-register-center/soul-register-client/soul-register-client-zookeeper/pom.xml
@@ -30,6 +30,11 @@
     <dependencies>
         <dependency>
             <groupId>org.dromara</groupId>
+            <artifactId>soul-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.dromara</groupId>
             <artifactId>soul-register-client-api</artifactId>
             <version>${project.version}</version>
         </dependency>
