+++ soul-register-center/soul-register-server/soul-register-server-zookeeper/src/main/java/org/dromara/soul/register/server/zookeeper/ZookeeperConfiguration.java
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.dromara.soul.register.server.zookeeper;
+
+import org.I0Itec.zkclient.ZkClient;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+
+/**
+ * ZookeeperConfiguration .
+ *
+ * @author lw1243925457
+ */
+@EnableConfigurationProperties(ZookeeperProperties.class)
+public class ZookeeperConfiguration {
+
+    /**
+     * register zkClient in spring ioc.
+     *
+     * @param zookeeperProp the zookeeper configuration
+     * @return ZkClient {@linkplain ZkClient}
+     */
+    @Bean
+    @ConditionalOnMissingBean(ZkClient.class)
+    public ZkClient zkClient(final ZookeeperProperties zookeeperProp) {
+        return new ZkClient(zookeeperProp.getUrl(), zookeeperProp.getSessionTimeout(), zookeeperProp.getConnectionTimeout());
+    }
+}
