+++ soul-spring-boot-starter/pom.xml
@@ -33,6 +33,7 @@
         <module>soul-spring-boot-starter-plugin</module>
         <module>soul-spring-boot-starter-sync-data-center</module>
         <module>soul-spring-boot-starter-client</module>
+        <module>soul-spring-boot-starter-register-center</module>
     </modules>
 
     <dependencies>
